move Escape class to Helper namespace
