+++ src/Helper/HtmlHelperMethods.php
@@ -3,8 +3,6 @@ declare(strict_types=1);
 
 namespace Qiq\Helper;
 
-use Qiq\Escape;
-
 trait HtmlHelperMethods
 {
     /**
